refactor(oidc): make OAuthGrantType the single source for grant types
Relocate OAuthGrantType from handlers/oidc/token.rs to a new
services/oidc/grant_type.rs (pub(crate)) so the discovery document can
derive grant_types_supported from it instead of a hand-maintained list
that could silently drift from the token endpoint's dispatch set.

Registration keeps ALLOWED_GRANT_TYPES as a documented superset (it adds
refresh_token, accepted but never issued), guarded by a new parity test
that every dispatched grant is registrable.

Closes #621
