+++ crates/vouch-server/src/handlers/oidc/tests/oidc_discovery.rs
@@ -125,10 +125,11 @@ async fn test_oidc_discovery_grant_types_match_token_parser() {
         .map(|v| v.as_str().expect("grant type should be string").to_string())
         .collect();
 
-    let parser_supported = crate::handlers::oidc::token::OAuthGrantType::supported_wire_values()
-        .into_iter()
-        .map(String::from)
-        .collect::<BTreeSet<_>>();
+    let parser_supported =
+        crate::services::oidc::grant_type::OAuthGrantType::supported_wire_values()
+            .into_iter()
+            .map(String::from)
+            .collect::<BTreeSet<_>>();
 
     assert_eq!(
         discovered, parser_supported,
@@ -136,7 +137,7 @@ async fn test_oidc_discovery_grant_types_match_token_parser() {
     );
 
     for grant in &discovered {
-        let parsed = grant.parse::<crate::handlers::oidc::token::OAuthGrantType>();
+        let parsed = grant.parse::<crate::services::oidc::grant_type::OAuthGrantType>();
         assert!(
             parsed.is_ok(),
             "discovery advertises unsupported grant type in parser: {grant}"
