+++ crates/vouch-server/src/handlers/oidc/token.rs
@@ -13,6 +13,7 @@ use crate::services::oidc::{
     ScopeSet,
     client_credentials::exchange_client_credentials,
     exchange::{TokenExchangeParams, exchange_token},
+    grant_type::{OAuthGrantType, ParseOAuthGrantTypeError},
     jwt_bearer::client_auth::{PendingJti, authenticate_client_jwt},
     token::{AuthCodeExchangeParams, exchange_authorization_code, validate_dpop_if_present},
 };
@@ -26,84 +27,6 @@ use axum::{
 use secrecy::SecretString;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
-
-/// OAuth grant types supported by this server.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum OAuthGrantType {
-    /// Standard OAuth 2.0 authorization code grant.
-    AuthorizationCode,
-    /// Client credentials grant (RFC 6749 Section 4.4).
-    ClientCredentials,
-    /// Device authorization grant (RFC 8628).
-    DeviceCode,
-    /// Token exchange grant (RFC 8693).
-    TokenExchange,
-    /// FIDO2 assertion grant (custom extension per RFC 6749 Section 4.5).
-    Fido2Assertion,
-}
-
-/// Parse error for OAuth `grant_type` values.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(super) struct ParseOAuthGrantTypeError {
-    value: String,
-}
-
-impl ParseOAuthGrantTypeError {
-    #[must_use]
-    pub(super) fn new(value: &str) -> Self {
-        Self {
-            value: value.to_string(),
-        }
-    }
-}
-
-impl std::fmt::Display for ParseOAuthGrantTypeError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "unsupported grant_type: {}", self.value)
-    }
-}
-
-impl std::error::Error for ParseOAuthGrantTypeError {}
-
-impl OAuthGrantType {
-    const SUPPORTED: [Self; 5] = [
-        Self::AuthorizationCode,
-        Self::ClientCredentials,
-        Self::DeviceCode,
-        Self::TokenExchange,
-        Self::Fido2Assertion,
-    ];
-
-    /// Wire-format `grant_type` value.
-    #[must_use]
-    pub(super) const fn as_str(self) -> &'static str {
-        match self {
-            Self::AuthorizationCode => "authorization_code",
-            Self::ClientCredentials => "client_credentials",
-            Self::DeviceCode => "urn:ietf:params:oauth:grant-type:device_code",
-            Self::TokenExchange => "urn:ietf:params:oauth:grant-type:token-exchange",
-            Self::Fido2Assertion => "urn:ietf:params:oauth:grant-type:fido2-assertion",
-        }
-    }
-
-    /// All supported `grant_type` wire values.
-    #[must_use]
-    pub(super) fn supported_wire_values() -> Vec<&'static str> {
-        Self::SUPPORTED.iter().copied().map(Self::as_str).collect()
-    }
-}
-
-impl std::str::FromStr for OAuthGrantType {
-    type Err = ParseOAuthGrantTypeError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Self::SUPPORTED
-            .iter()
-            .copied()
-            .find(|grant_type| grant_type.as_str() == s)
-            .ok_or_else(|| ParseOAuthGrantTypeError::new(s))
-    }
-}
 
 /// Token response (RFC 6749 Section 5.1).
 #[derive(Serialize)]
@@ -1364,59 +1287,4 @@ fn token_error_response(error: &str, description: &str) -> Response {
         }),
     )
         .into_response()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::OAuthGrantType;
-
-    #[test]
-    fn test_oauth_grant_type_from_str_authorization_code() {
-        let result: Result<OAuthGrantType, _> = "authorization_code".parse();
-        assert_eq!(result, Ok(OAuthGrantType::AuthorizationCode));
-    }
-
-    #[test]
-    fn test_oauth_grant_type_from_str_device_code() {
-        let result: Result<OAuthGrantType, _> =
-            "urn:ietf:params:oauth:grant-type:device_code".parse();
-        assert_eq!(result, Ok(OAuthGrantType::DeviceCode));
-    }
-
-    #[test]
-    fn test_oauth_grant_type_from_str_token_exchange() {
-        let result: Result<OAuthGrantType, _> =
-            "urn:ietf:params:oauth:grant-type:token-exchange".parse();
-        assert_eq!(result, Ok(OAuthGrantType::TokenExchange));
-    }
-
-    #[test]
-    fn test_oauth_grant_type_from_str_fido2_assertion() {
-        let result: Result<OAuthGrantType, _> =
-            "urn:ietf:params:oauth:grant-type:fido2-assertion".parse();
-        assert_eq!(result, Ok(OAuthGrantType::Fido2Assertion));
-    }
-
-    #[test]
-    fn test_oauth_grant_type_from_str_client_credentials() {
-        let result: Result<OAuthGrantType, _> = "client_credentials".parse();
-        assert_eq!(result, Ok(OAuthGrantType::ClientCredentials));
-    }
-
-    #[test]
-    fn test_oauth_grant_type_from_str_rejects_unknown() {
-        let result: Result<OAuthGrantType, _> = "password".parse();
-        assert!(result.is_err());
-
-        let result2: Result<OAuthGrantType, _> = "".parse();
-        assert!(result2.is_err());
-
-        let result3: Result<OAuthGrantType, _> = "jwt-bearer".parse();
-        assert!(result3.is_err());
-
-        // Lock-in: §2.1 grant URN must be rejected (RFC 7523 §2.1 removed).
-        let bearer: Result<OAuthGrantType, _> =
-            "urn:ietf:params:oauth:grant-type:jwt-bearer".parse();
-        assert!(bearer.is_err());
-    }
 }
